Port Craft.NestedElementManager to a modern module
Moves the legacy jQuery NestedElementManager into
resources/js/modules/nested-element-manager/ following the shared module
pattern: a logic class on @craftcms/garnish Base (modern Select/DragSort
for card selection and drag-sorting, documented jQuery seams for the
Craft.ui buttons, disclosure/expandable menus, element index, editor
slideouts, and Craft.cp), a <craft-nested-element-manager>
ControllerElement that boots from an element-type attribute and a JSON
settings attribute, a WeakMap instance registry, and an index shim that
assigns window.Craft.NestedElementManager (the constructor still accepts
the legacy three-argument signature for plugin boots).

NestedElementManager::createView() now emits the custom element around
its rendered markup instead of an HtmlStack JS boot, so the manager
self-boots and re-boots across Inertia fragment swaps. destroy() now
actually tears down listeners and the Select/DragSort instances.
