+++ packages/craftcms-legacy/cp/src/Craft.js
@@ -73,7 +73,6 @@ import './js/LightSwitch.js';
 import './js/LinkField.js';
 import './js/LinkInput.js';
 import './js/LivePreview.js';
-import './js/NestedElementManager.js';
 import './js/PasswordInput.js';
 import './js/Preview.js';
 import './js/PreviewFileModal.js';
