+++ resources/js/cp.ts
@@ -19,6 +19,7 @@ import './modules/grouped-entry-type-manager';
 import './modules/queue';
 import './modules/slideout';
 import './modules/auth-method-setup';
+import './modules/nested-element-manager';
 
 window.Cp = Cp as unknown as typeof window.Cp;
