+++ resources/js/legacy.ts
@@ -18,6 +18,7 @@ import './modules/grouped-entry-type-manager/index';
 import './modules/queue/index';
 import './modules/slideout/index';
 import './modules/auth-method-setup/index';
+import './modules/nested-element-manager/index';
 
 const {default: Cp} = await import('./bootstrap/cp.js');
